BUGFIX: Relax condition for inline/sprite icon
Currently, `inline = null` leads to that no icon get rendered at all. The `inline` condition is simply set to `true`, at there is no need for further checks. If someone wants to integrate another way to show the icon (for example with an included sprite on the page, it is possible to prepend it with the `@position` property.